Move file content reading to a new method
Simple refactor, moving the repeated code where I read content of
a file into a scalar. The unsetting/setting of "$/" I missed once
whilst writing code; so although this is a small change it should
mean that I don't make that mistake again... and decrease the lines
of code too! :-)